+++ sdk/openai/azure-ai-openai/src/main/java/com/azure/ai/openai/OpenAIAsyncClient.java
@@ -904,7 +904,6 @@ public final class OpenAIAsyncClient {
         ChatCompletionsOptions chatCompletionsOptions, ChatCompletionStreamOptions streamOptions) {
         ChatCompletionsOptionsAccessHelper.setStream(chatCompletionsOptions, true);
         ChatCompletionsOptionsAccessHelper.setStreamOptions(chatCompletionsOptions, streamOptions);
-
         RequestOptions requestOptions = new RequestOptions();
         Flux<ByteBuffer> responseStream
             = getChatCompletionsWithResponse(deploymentOrModelName, BinaryData.fromObject(chatCompletionsOptions),